+++ packages/utils/src/misc.ts
@@ -24,3 +24,13 @@ export function arrayToEnum<
 export function includesGuard<T>(ls: readonly T[], v: unknown): v is T {
   return ls.includes(v as T);
 }
+
+/** silence "error TSTS2322" but it should be "practically" safer than `any` */
+export function saferFunctionCast<F extends (...args: any[]) => any>(
+  f: (...args: Parameters<F>) => ReturnType<F>
+): F {
+  // silence type error:
+  //   error TS2322: Type '(...args: Parameters<F>) => ReturnType<F>' is not assignable to type 'F'.
+  //   '(...args: Parameters<F>) => ReturnType<F>' is assignable to the constraint of type 'F', but 'F' could be instantiated with a different subtype of constraint '(...args: any[]) => any'.
+  return f as any;
+}
